MK-572 - added edition and placed restrictions on viewing the permiss…

+++ mica-rest/src/main/java/org/obiba/mica/security/rest/SubjectAclResource.java
@@ -48,6 +48,7 @@ public class SubjectAclResource {
 
   @GET
   public List<AclDto> get() {
+    subjectAclService.checkPermission(resource, "EDIT", instance);
     return subjectAclService.findByResourceInstance(resource, instance).stream().map(
       a -> AclDto.newBuilder().setType(a.getType().name()).setPrincipal(a.getPrincipal()).setResource(resource)
         .setRole(PermissionsUtils.asRole(a.getActions())).setInstance(instance).build()).collect(Collectors.toList());
@@ -55,6 +56,7 @@ public class SubjectAclResource {
 
   @DELETE
   public Response delete(@QueryParam("principal") String principal, @QueryParam("type") String typeStr) {
+    subjectAclService.checkPermission(resource, "EDIT", instance);
     SubjectAcl.Type type = SubjectAcl.Type.valueOf(typeStr.toUpperCase());
     subjectAclService.removeSubjectPermissions(type, principal, resource, instance);
     subjectAclService.removeSubjectPermissions(type, principal, fileResource, fileInstance);
@@ -64,6 +66,7 @@ public class SubjectAclResource {
   @PUT
   public Response update(@QueryParam("principal") String principal, @QueryParam("type") String typeStr,
     @QueryParam("role") String role) {
+    subjectAclService.checkPermission(resource, "EDIT", instance);
     SubjectAcl.Type type = SubjectAcl.Type.valueOf(typeStr.toUpperCase());
     String actions = PermissionsUtils.asActions(role.toUpperCase());
     subjectAclService.addSubjectPermission(type, principal, resource, actions, instance);
